+++ src/main/kotlin/cl/ravenhill/keen/limits/GenerationCount.kt
@@ -17,7 +17,7 @@ import cl.ravenhill.jakt.constraints.ints.BePositive
  * @version 2.0.0
  * @since 1.0.0
  */
-data class GenerationCount(private val i: Int) : Match({ generation >= i }) {
+data class GenerationCount(private val i: Int) : MatchLimit({ generation >= i }) {
     init {
         constraints { "Generation count [$i] must be at least 1" { i must BePositive } }
     }