+++ src/main/kotlin/cl/ravenhill/keen/limits/SteadyGenerations.kt
@@ -9,7 +9,7 @@ import cl.ravenhill.jakt.Jakt.constraints
 import cl.ravenhill.jakt.constraints.ints.BePositive
 
 /**
- * A [Match] limit that checks if the population has remained steady for a given number of
+ * A [MatchLimit] limit that checks if the population has remained steady for a given number of
  * generations.
  *
  * @property n The number of steady generations required to satisfy the limit.
@@ -18,7 +18,7 @@ import cl.ravenhill.jakt.constraints.ints.BePositive
  * @since 1.0.0
  * @version 2.0.0
  */
-data class SteadyGenerations(val n: Int) : Match({ steadyGenerations >= n }) {
+data class SteadyGenerations(val n: Int) : MatchLimit({ steadyGenerations >= n }) {
     init {
         constraints { "Steady generations must be positive" { n must BePositive } }
     }